modularized CPU side terrain math
